Move dataset testing utils
