+++ doc/conftest.py
@@ -1,1 +1,4 @@
-from nilearn.datasets._testing import request_mocker, temp_nilearn_data_dir
+from nilearn.datasets.tests._testing import (
+    request_mocker,
+    temp_nilearn_data_dir,
+)
